Se completa la funcion de regresion lineal con GSL - se mejora la implementacion del ordenamiento paralelo - se implementa lectura paralela de archivo

+++ secuencial/test/manejador_de_archivos_secuencial_test.cc
@@ -27,7 +27,7 @@ SUITE(ManejadorDeArchivosSecuencialTest)
             ifstream lectura (archivo);
             REQUIRE CHECK(lectura.is_open());
             string linea;
-            for(unsigned int i=0;i<esperado.size();i++)
+            for(size_t i=0;i<esperado.size();i++)
             {
                 CHECK(getline(lectura, linea));
                 CHECK_EQUAL(esperado[i],linea);
